requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
psycopg2-binary==2.9.10
pydantic==2.11.10
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
